backend/requirements.txt
SQLAlchemy==2.0.32
psycopg[binary]==3.2.*
alembic==1.13.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
pydantic==2.8.2
email-validator==2.1.1
python-multipart==0.0.9
httpx==0.27.0

prometheus-client==0.20.0

pytz>=2023.3
